opa: remove "omitted" and "empty array" cases for input.projects
You might argue that we need them; but this is quite in line with what
happens for the other inputs: if you don't provide them, you're unlikely
to get a meaningful result.

Since we control the input (OPA's `input` is nowhere near a public
interface), we should get by with this.

Spot-checking partial results, we find that this gets us rid of 2/3 of the checks:

    $ opa run -w authz_v2.rego common.rego ../example_v2/store-pretty.json
    OPA 0.10.8-dev (commit 0ed56fef-dirty, built at 2019-05-13T08:29:47Z)

    Run 'help' to see a list of commands.

    > unknown input
    > data.authz_v2.allowed_project[x]

...yields 64 queries and one support rule now; without this change, it would be 192 queries.
`denied_project` would add another 78 (26 now).

I think there's further simplifications to be made... let's see.

Signed-off-by: Stephan Renatus <srenatus@chef.io>